Major Modification of the Graphene Fano project.
The code here was copied from the Version 3, and will
be modified to account for the site-resolved impurity averaging.

Note that the code on my local machine was bjorked because
of funky TABS turned to SPACE (probably a vi problem). I have
tried to remove all TABS and keep the proper indentation.
Computation with this version of the code (which should be exactly
the same as V3 at this point) yielded exactly the same data as
the V3 code for the test run
HCF_nq=4_8_1_nk=4_8_1_mu=-400_meV_Gamma=50_meV.nc

(This is no proof that the code is correct, and in a better
world I would implement unit tests to make sure everything is
100% working).