A new IC class has been added.
This class implements the computation of IC, the
direct matsubara sum over the gammaC kernel, which
is the "real" kernel, minus its behavior at infinity
and a padé expansion about infinity. It is expected that
GammaC will be wild and funky near energy zero, but will
decay very fast as |z| -> infinity.